Fix build with single-header simdjson installations
