+++ libvast/builtins/formats/json.cpp
@@ -32,6 +32,56 @@ namespace vast::plugins::json {
 
 namespace {
 
+/// A variant of *to_lines* that returns a string view with additional padding
+/// bytes that are safe to read.
+inline auto to_padded_lines(generator<chunk_ptr> input)
+  -> generator<std::optional<simdjson::padded_string_view>> {
+  auto buffer = std::string{};
+  bool ended_on_linefeed = false;
+  for (auto&& chunk : input) {
+    if (!chunk || chunk->size() == 0) {
+      co_yield std::nullopt;
+      continue;
+    }
+    const auto* begin = reinterpret_cast<const char*>(chunk->data());
+    const auto* const end = begin + chunk->size();
+    if (ended_on_linefeed && *begin == '\n') {
+      ++begin;
+    };
+    ended_on_linefeed = false;
+    for (const auto* current = begin; current != end; ++current) {
+      if (*current != '\n' && *current != '\r') {
+        continue;
+      }
+      const auto capacity = static_cast<size_t>(end - begin);
+      const auto size = static_cast<size_t>(current - begin);
+      if (buffer.empty() and capacity >= size + simdjson::SIMDJSON_PADDING) {
+        co_yield simdjson::padded_string_view{begin, size, capacity};
+      } else {
+        buffer.append(begin, current);
+        buffer.reserve(buffer.size() + simdjson::SIMDJSON_PADDING);
+        co_yield simdjson::padded_string_view{buffer};
+        buffer.clear();
+      }
+      if (*current == '\r') {
+        auto next = current + 1;
+        if (next == end) {
+          ended_on_linefeed = true;
+        } else if (*next == '\n') {
+          ++current;
+        }
+      }
+      begin = current + 1;
+    }
+    buffer.append(begin, end);
+    co_yield std::nullopt;
+  }
+  if (!buffer.empty()) {
+    buffer.reserve(buffer.size() + simdjson::SIMDJSON_PADDING);
+    co_yield simdjson::padded_string_view{buffer};
+  }
+}
+
 enum class parser_action {
   skip = 0,
   yield = 1,
